Seperate reducers
